changed dbrc utils
