+++ tests/functional/module_utils/ims_test_dbrc_utils.py
@@ -1,0 +1,17 @@
+class DBRCInputParameters():
+    HLQ1 = "IMSTESTL."
+    HLQ2 = "IMS1."
+    STEPLIB = [
+        "IMSTESTU.IMS1501.MARKER",
+        HLQ1 + HLQ2 + "EXITLIB",
+        "IMSTESTG.IMS15R.TSTRES",
+        "IMSBLD.IMS15R.USERLIB",
+        "IMSBLD.I15RTSMM.CRESLIB"
+    ]
+    DBDLIB = HLQ1 + HLQ2 + "DBDLIB"
+    DYNALLOC = HLQ1 + HLQ2 + "SDFSRESL"
+    GENJCL = HLQ1 + HLQ2 + "GENJCL"
+    JCLOUT = HLQ1 + HLQ2 + "JCLOUT"
+    RECON1 = HLQ1 + HLQ2 + "RECON1"
+    RECON2 = HLQ1 + HLQ2 + "RECON2"
+    RECON3 = HLQ1 + HLQ2 + "RECON3"
