+++ tests/functional/module_utils/ims_test_gen_utils.py
@@ -71,19 +71,3 @@ class ACBInputParameters():
     INVALID_PDSE = "IMSTESTL.ANS.PDSE"
     EMPTY_PSBLIB = "IMSTESTL.ANS.EMPTY.PSBLIB"
     EMPTY_DBDLIB = "IMSTESTL.ANS.EMPTY.DBDLIB"
-
-class DBRCInputParameters():
-    STEPLIB = [
-        "IMSTESTU.IMS1501.MARKER",
-        "IMSBANK2.IMS1.EXITLIB",
-        "IMSTESTG.IMS15R.TSTRES",
-        "IMSBLD.IMS15R.USERLIB",
-        "IMSBLD.I15RTSMM.CRESLIB"
-    ]
-    DBDLIB = "IMSBANK2.IMS1.DBDLIB"
-    DYNALLOC = "IMSBANK2.IMS1.SDFSRESL"
-    GENJCL = "IMSTESTL.IMS1.GENJCL"
-    JCLOUT = "IMSBANK2.IMS1.JCLOUT"
-    RECON1 = "IMSBANK2.IMS1.RECON1"
-    RECON2 = "IMSBANK2.IMS1.RECON2"
-    RECON3 = "IMSBANK2.IMS1.RECON3"
