+++ tests/functional/modules/ims_dbrc/test_ims_dbrc_sample.py
@@ -2,7 +2,7 @@
 
 from pprint import pprint
 import pytest
-from ibm_zos_ims.tests.functional.module_utils.ims_test_gen_utils import DBRCInputParameters as ip # pylint: disable=import-error
+from ibm_zos_ims.tests.functional.module_utils.ims_test_dbrc_utils import DBRCInputParameters as ip # pylint: disable=import-error
 from ansible_collections.ibm.ibm_zos_ims.plugins.module_utils.ims_module_error_messages import DBRCErrorMessages as em # pylint: disable=import-error
 __metaclass__ = type
